added, component PokemonList to reduce complexity in App and be more modular
